refactor(ui): Rename plain page files to `index.tsx`
The route directory includes a mix of page files named `index.tsx` and
files with the last part of the URL in the file name, like
`create-organiozation.tsx`. Use `directory/index.tsx` for all
pages for consistency.
