+++ ui/src/routes/organizations/$orgId/edit/index.tsx
@@ -18,13 +18,17 @@
  */
 
 import { zodResolver } from '@hookform/resolvers/zod';
+import { useSuspenseQuery } from '@tanstack/react-query';
 import { createFileRoute, useNavigate } from '@tanstack/react-router';
 import { Loader2 } from 'lucide-react';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
-import { useSecretsServicePostSecretForOrganization } from '@/api/queries';
-import { ApiError } from '@/api/requests';
+import {
+  useOrganizationsServiceGetOrganizationByIdKey,
+  useOrganizationsServicePatchOrganizationById,
+} from '@/api/queries';
+import { ApiError, OrganizationsService } from '@/api/requests';
 import { ToastError } from '@/components/toast-error';
 import { Button } from '@/components/ui/button';
 import {
@@ -46,22 +50,29 @@ import { toast } from '@/lib/toast';
 
 const formSchema = z.object({
   name: z.string(),
-  value: z.string(),
   description: z.string().optional(),
 });
 
-const CreateOrganizationSecretPage = () => {
-  const navigate = useNavigate();
+const EditOrganizationPage = () => {
   const params = Route.useParams();
+  const navigate = useNavigate();
 
-  const { mutateAsync, isPending } = useSecretsServicePostSecretForOrganization(
-    {
+  const { data: organization } = useSuspenseQuery({
+    queryKey: [useOrganizationsServiceGetOrganizationByIdKey, params.orgId],
+    queryFn: async () =>
+      await OrganizationsService.getOrganizationById({
+        organizationId: Number.parseInt(params.orgId),
+      }),
+  });
+
+  const { mutateAsync, isPending } =
+    useOrganizationsServicePatchOrganizationById({
       onSuccess(data) {
-        toast.info('Create Organization Secret', {
-          description: `New organization secret "${data.name}" created successfully.`,
+        toast.info('Edit Organization', {
+          description: `Organization "${data.name}" updated successfully.`,
         });
         navigate({
-          to: '/organizations/$orgId/secrets',
+          to: '/organizations/$orgId',
           params: { orgId: params.orgId },
         });
       },
@@ -75,19 +86,21 @@ const CreateOrganizationSecretPage = () => {
           },
         });
       },
-    }
-  );
+    });
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
+    defaultValues: {
+      name: organization.name,
+      description: organization.description ?? undefined,
+    },
   });
 
   async function onSubmit(values: z.infer<typeof formSchema>) {
     await mutateAsync({
-      organizationId: Number.parseInt(params.orgId),
+      organizationId: organization.id,
       requestBody: {
         name: values.name,
-        value: values.value,
         description: values.description,
       },
     });
@@ -95,7 +108,7 @@ const CreateOrganizationSecretPage = () => {
 
   return (
     <Card>
-      <CardHeader>Create Organization Secret</CardHeader>
+      <CardHeader>Edit Organization</CardHeader>
       <Form {...form}>
         <form onSubmit={form.handleSubmit(onSubmit)} className='space-y-8'>
           <CardContent className='space-y-4'>
@@ -106,19 +119,6 @@ const CreateOrganizationSecretPage = () => {
                 <FormItem>
                   <FormLabel>Name</FormLabel>
                   <FormControl autoFocus>
-                    <Input {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name='value'
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Value</FormLabel>
-                  <FormControl>
                     <Input {...field} />
                   </FormControl>
                   <FormMessage />
@@ -140,16 +140,23 @@ const CreateOrganizationSecretPage = () => {
             />
           </CardContent>
           <CardFooter>
+            <Button
+              type='button'
+              className='m-1'
+              variant='outline'
+              onClick={() => navigate({ to: '/organizations/' + params.orgId })}
+              disabled={isPending}
+            >
+              Cancel
+            </Button>
             <Button type='submit' disabled={isPending}>
               {isPending ? (
                 <>
-                  <span className='sr-only'>
-                    Creating organization secret...
-                  </span>
+                  <span className='sr-only'>Editing organization...</span>
                   <Loader2 size={16} className='mx-3 animate-spin' />
                 </>
               ) : (
-                'Create'
+                'Submit'
               )}
             </Button>
           </CardFooter>
@@ -159,8 +166,15 @@ const CreateOrganizationSecretPage = () => {
   );
 };
 
-export const Route = createFileRoute(
-  '/organizations/$orgId/secrets/create-secret'
-)({
-  component: CreateOrganizationSecretPage,
+export const Route = createFileRoute('/organizations/$orgId/edit/')({
+  loader: async ({ context, params }) => {
+    await context.queryClient.ensureQueryData({
+      queryKey: [useOrganizationsServiceGetOrganizationByIdKey, params.orgId],
+      queryFn: () =>
+        OrganizationsService.getOrganizationById({
+          organizationId: Number.parseInt(params.orgId),
+        }),
+    });
+  },
+  component: EditOrganizationPage,
 });
