+++ ui/src/routes/organizations/$orgId/products/$productId/repositories/$repoId/-components/package-manager-field.tsx
@@ -44,7 +44,7 @@ import { packageManagers } from '@/routes/organizations/$orgId/products/$product
 import {
   CreateRunFormValues,
   PackageManagerId,
-} from '../_repo-layout/-create-run-utils';
+} from '../_repo-layout/create-run/-create-run-utils';
 
 type PackageManagerFieldProps = {
   form: UseFormReturn<CreateRunFormValues>;
