+++ ui/src/routes/organizations/$orgId/products/$productId/repositories/$repoId/-components/reporter-fields.tsx
@@ -28,7 +28,7 @@ import {
 import { FormControl, FormField } from '@/components/ui/form';
 import { Switch } from '@/components/ui/switch';
 import { reportFormats } from '@/routes/organizations/$orgId/products/$productId/repositories/$repoId/-types';
-import { CreateRunFormValues } from '../_repo-layout/-create-run-utils';
+import { CreateRunFormValues } from '../_repo-layout/create-run/-create-run-utils';
 
 type ReporterFieldsProps = {
   form: UseFormReturn<CreateRunFormValues>;
